+++ tests/pricing_models/test_calc_out_given_in.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import decimal
 import unittest
 import numpy as np
-from test_dataclasses import (
+from dataclass_testcases import (
     TestCaseCalcOutGivenInSuccess,
     TestResultCalcOutGivenInSuccess,
     TestResultCalcOutGivenInSuccessByModel,
